feat: expose utils module
