+++ package.json
@@ -7,23 +7,28 @@
     "README.md"
   ],
   "exports": {
+    "./utils": "./dist/utils.js",
     "./*": "./dist/*.json"
   },
   "main": "./dist/versions.json",
   "module": "./dist/versions.json",
   "license": "GPL-3.0-only",
   "engines": {
-    "node": "^16 || ^18 || ^20",
-    "npm": "^8 || ^9 || ^10"
+    "node": "^18 || ^20 || ^22",
+    "npm": "^10"
   },
   "scripts": {
-    "build": "ts-node scripts/extract-contract-abis.ts",
+    "build": "ts-node scripts/extract-contract-abis.ts && tsc lib/utils.ts --skipLibCheck --declaration --outDir dist",
     "test": "node scripts/check-versions.mjs",
     "prepare": "husky install"
   },
+  "dependencies": {
+    "abitype": "^1.0.7",
+    "lodash": "^4.17.21"
+  },
   "devDependencies": {
-    "@commitlint/cli": "^19.3.0",
-    "@commitlint/config-conventional": "^19.2.2",
+    "@commitlint/cli": "^19.6.1",
+    "@commitlint/config-conventional": "^19.6.0",
     "@types/lodash": "^4.14.191",
     "@types/mkdirp": "^1.0.2",
     "@types/rimraf": "^3.0.2",
@@ -39,12 +44,11 @@
     "eslint-plugin-prettier": "^4.2.1",
     "fast-glob": "^3.2.11",
     "husky": "^8.0.3",
-    "lodash": "^4.17.21",
     "mkdirp": "^1.0.4",
     "prettier": "^2.8.0",
     "rimraf": "^3.0.2",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.3",
+    "typescript": "^5.7.2",
     "yargs": "^17.3.1"
   }
 }
